chore(api): split test_api.py (~743 LOC) into per-feature files
The monolithic test file mixed route-handler tests, recommendations
logic tests and the digest orchestration tests. Splitting keeps each
file focused on one collaborator and matches the pattern auto-bid
already follows (`test_auto_bid.py`).

- `test_routes.py` — Flask handlers: /health, /version, /scraper/trigger,
  /digests/daily, /market/auto-bid, /teams, /managers, /market,
  /lineups/auto-pick, /budget/recommendations. Pins wiring + 500s only.
- `test_recommendations.py` — `recommendations.py` unit tests:
  `compute_dynamic_margin`, `_filter_affordable` (including the sole-GK
  house rule), `_gather_rivals`, `_pick_top_per_position`,
  `_format_telegram_text`.
- `test_digests.py` — `digests.run_daily` orchestration:
  credential-missing short-circuit, auto-bid chaining,
  swallow-on-auto-bid-failure.

Same test names + assertions throughout — zero coverage delta. The
deleted `test_api.py` was 743 LOC; the three new files are 380 + 215 +
135 = 730 LOC total (the slight drop is from dropping the duplicated
`_row` helper now defined once in `test_recommendations.py`).
